+++ ORBITER/src/lib/group_actions/actions/interface_permutation_representation.cpp
@@ -24,7 +24,7 @@ namespace group_actions {
 // #############################################################################
 
 
-static int permutation_representation_group_element_image_of(action &A, int a,
+static long int permutation_representation_group_element_image_of(action &A, long int a,
 	void *elt, int verbose_level);
 static void permutation_representation_group_element_image_of_low_level(action &A,
 	int *input, int *output, void *elt, int verbose_level);
@@ -69,12 +69,12 @@ static void permutation_representation_group_element_print_latex(action &A,
 static void permutation_representation_group_element_print_latex_with_print_point_function(
 	action &A,
 	void *elt, std::ostream &ost,
-	void (*point_label)(std::stringstream &sstr, int pt, void *data),
+	void (*point_label)(std::stringstream &sstr, long int pt, void *data),
 	void *point_label_data);
 static void permutation_representation_group_element_print_verbose(action &A,
 	void *elt, std::ostream &ost);
 static void permutation_representation_group_print_point(action &A,
-	int a, std::ostream &ost);
+	long int a, std::ostream &ost);
 
 
 void action_pointer_table::init_function_pointers_permutation_representation_group()
@@ -113,13 +113,13 @@ void action_pointer_table::init_function_pointers_permutation_representation_gro
 
 
 
-static int permutation_representation_group_element_image_of(action &A,
-		int a, void *elt, int verbose_level)
+static long int permutation_representation_group_element_image_of(action &A,
+		long int a, void *elt, int verbose_level)
 {
 	int f_v = (verbose_level >= 1);
 	permutation_representation &P = *A.G.Permutation_representation;
 	int *Elt = (int *) elt;
-	int b;
+	long int b;
 
 	if (f_v) {
 		cout << "permutation_representation_group_element_image_of "
@@ -441,7 +441,7 @@ static void permutation_representation_group_element_print_latex(
 static void permutation_representation_group_element_print_latex_with_print_point_function(
 	action &A,
 	void *elt, std::ostream &ost,
-	void (*point_label)(std::stringstream &sstr, int pt, void *data),
+	void (*point_label)(std::stringstream &sstr, long int pt, void *data),
 	void *point_label_data)
 {
 	cout << "permutation_representation_group_element_print_latex_with_print_point_function "
@@ -459,7 +459,7 @@ static void permutation_representation_group_element_print_verbose(
 
 }
 
-static void permutation_representation_group_print_point(action &A, int a, ostream &ost)
+static void permutation_representation_group_print_point(action &A, long int a, ostream &ost)
 {
 	//permutation_representation &P = *A.G.Permutation_representation;
